feat: changed tenant provider
